chore: move mapToHex to utils and move categories out of scope to avoid recreates
